requirements.txt
asgiref==3.5.2
attrs==22.1.0
backports.zoneinfo==0.2.1
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-coverage==1.2.4
django-heroku==0.3.1
django-markdown-deux==1.0.5
django-nose==1.4.7
django-on-heroku==1.1.2
django-rest-framework==0.1.0
django-rest-knox==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
markdown2==2.4.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
nose==1.3.7
oauthlib==3.2.1
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.3
pytest-cov==4.0.0
pytest-django==4.5.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
validate-email==1.3
whitenoise==6.2.0
zipp==3.8.1
